requirements.txt
yt_dlp
python-telegram-bot
httpx
anyio
certifi
h11
httpcore
sniffio